+++ timetracker/cfg/utils.py
@@ -12,11 +12,26 @@ from os.path import abspath
 from os.path import normpath
 from os.path import dirname
 from os.path import join
+from os.path import ismount
 ##from os.path import commonpath
 from os.path import commonprefix
 from logging import debug
 from logging import warning
 from tomlkit import parse
+
+
+def finddirtrk(path, trksubdir):
+    """Walk up dirs until find .timetracker/ proj dir or mount dir"""
+    path = abspath(path)
+    trkdir = join(path, trksubdir)
+    if exists(trkdir):
+        return trkdir, True
+    while not ismount(path):
+        trkdir = join(path, trksubdir)
+        if exists(trkdir):
+            return trkdir, True
+        path = dirname(path)
+    return path, False
 
 def parse_cfg(fin, desc=''):
     """Read a config file and load it into a TOML document"""
@@ -32,12 +47,14 @@ def _read_local_cfg(fin):
 
 def get_relpath_adj(projdir, dirhome):
     """Collapse an absolute pathname into one with a `~` if projdir is a child of home"""
-    if has_homedir(dirhome, projdir):
+    if has_homedir(abspath(dirhome), projdir):
         return join('~', relpath(projdir, dirhome))
     return projdir
 
 def has_homedir(homedir, projdir):
     """Checks to see if `projdir` has a root of `rootdir`"""
+    assert homedir == abspath(homedir)
+    assert projdir == abspath(projdir)
     homedir = abspath(homedir)
     ##debug(f'has_homedir      homedir {homedir}')
     ##debug(f'has_homedir      projdir {projdir}')
@@ -65,7 +82,7 @@ def replace_envvar(fpat):
 
 def get_filename_abs(fname):
     """Get the absolute filename"""
-    return normpath(abspath(expanduser(fname)))
+    return abspath(expanduser(fname))
 
 def get_dirname_abs(fname):
     """Get the absolute directory name"""
@@ -82,6 +99,8 @@ def chk_isdir(dname, prefix=''):
 
 def replace_homepath(fname):
     """Replace '~' with the expanded filepath"""
+    # pylint: disable=fixme
+    # TODO: use commonprefix
     fname = normpath(fname)
     home_str = expanduser('~')
     home_len = len(home_str)
@@ -110,6 +129,14 @@ def get_dirhome(dirhome):
             return ret
         raise RuntimeError(f'NO DIRECTORY IN ENVVAR {dirhome}: {dirhome}')
     raise RuntimeError(f'UNKNOWN DIRECTORY FOR CONFIGURATION: {dirhome}')
+
+def get_shortest_name(filename):
+    """Return the shortest filename"""
+    fabs = abspath(filename)
+    # pylint: disable=fixme
+    # TODO: use commonprefix
+    frel = normpath(relpath(filename))
+    return fabs if len(fabs) < len(frel) else frel
 
 
 # Copyright (C) 2025-present, DV Klopfenstein, PhD. All rights reserved.