+++ timetracker/trk.py
@@ -3,29 +3,35 @@
 __copyright__ = 'Copyright (C) 2025-present, DV Klopfenstein, PhD. All rights reserved.'
 __author__ = "DV Klopfenstein, PhD"
 
+from os import getcwd
 from os.path import exists
-from logging import error
-#from logging import DEBUG
+#from logging import error
+
 #from logging import basicConfig
+#from logging import DEBUG
+###from logging import INFO
+
 from timetracker.filemgr import FileMgr
 from timetracker.cfg.cfg_global import CfgGlobal
 from timetracker.cfg.cfg_local import CfgProj
 from timetracker.cli import Cli
+from timetracker.msgs import str_started
 from timetracker.cmd.init import run_init
 from timetracker.cmd.start import run_start
 from timetracker.cmd.stop import run_stop
-from timetracker.msgs import prt_started
+from timetracker.cmd.csvupdate import run_csvupdate
 
 fncs = {
     'init': run_init,
     'start': run_start,
     'stop': run_stop,
+    'csvupdate': run_csvupdate,
 }
 
 
 def main():
     """Connect all parts of the timetracker"""
-    ##basicConfig(level=DEBUG)
+    #basicConfig(level=DEBUG)
     obj = TimeTracker()
     obj.run()
 
@@ -59,12 +65,12 @@ class TimeTracker:
             print('Run `trk start` to begin timetracking')
         else:
             self.cfg_local.prt_elapsed()
-            prt_started()
+            print(str_started())
 
     def _msg_init(self):
         self.cli.parser.print_help()
-        print('')
-        error('Run `trk init` to initialize local timetracker')
+        print('\nRun `trk init` to initialize time-tracking '
+              f'for the project in {getcwd()}')
 
 
 # Copyright (C) 2025-present, DV Klopfenstein, PhD. All rights reserved.